requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
Django==4.0.2
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pylint==2.12.2
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
wrapt==1.13.3
django-mptt==0.13.4
pre-commit==2.17.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
pillow==9.0.1